created function for getting user data and refactored other bits of code

+++ database/test2.py
@@ -37,12 +37,11 @@ def instructions() -> None:
     print("UPDATE : update/ change existing data in the database.")
     print("DELETE : delete data in the database.")
     print("QUIT : close program.")
-    print("INSTRUCTIONS : view these instructions again.\n")
+    print("INSTRUCTIONS : view these instructions again.\n\n")
 
 
 def get_user_key() -> str:
-    key = input("\nEnter key: ")
-    print()
+    key = input("Enter key: ")
     key = key.upper()
 
     return key
@@ -62,29 +61,54 @@ def execute_operations(key:str, session) -> None:
     elif key == "QUIT":
         sys.exit()
     else:
-        print("Not a key.\n")
+        print("Not a key.\n\n")
+
+
+def get_data(session) -> list:
+    data = []
+
+    new_first_name = input("Enter first name: ")
+    new_first_name = new_first_name.capitalize()
+    data.append(new_first_name)
+
+    new_last_name = input("Enter last name: ")
+    new_last_name = new_last_name.capitalize()
+    data.append(new_last_name)
+
+    while True:
+        new_dob = input("Enter date of birth (yyyy-mm-dd): ")
+        if len(new_dob) != 10:
+            print("Enter an acceptable date of birth (yyyy/mm/dd).")
+        else:
+            break
+        
+    data.append(new_dob)
+
+    return data
 
 
 def create(session:None) -> None:
-    new_first_name = input("Enter first name: ")
-    new_first_name = new_first_name.capitalize()
-    new_last_name = input("Enter last name: ")
-    new_last_name = new_last_name.capitalize()
-    new_dob = input("Enter date of birth (yyyy-mm-dd): ")
+    create_data = get_data(session)
+
+
+    new_first_name, new_last_name, new_dob = create_data[0], create_data[1], create_data[2]
 
     people = session.query(People)
 
     for person in people:
         if new_first_name == person.first_name and new_last_name == person.last_name:
-            print("Person already in database.")
+            print(f"{new_first_name} {new_last_name} already in database.")
+            print("\n")
             return
         else:
             new_person = People(first_name = new_first_name, last_name = new_last_name, dob = new_dob)
+            print("\n")
             session.add(new_person)
             session.commit()
             return 
 
     new_person = People(first_name = new_first_name, last_name = new_last_name, dob = new_dob)
+    print("\n")
     session.add(new_person)
     session.commit()
     return 
@@ -95,53 +119,42 @@ def read(session:None) -> None:
     for person in people:
         print(person.first_name, person.last_name, person.dob)
         
-    print()
+    print("\n")
 
 
 def update(session:None) -> None:
     print("To update a persons data, start by entering the first and last name of the person you wish to update.")
-    print()
 
-    prev_first_name = input("Enter previous first name: ")
-    prev_first_name = prev_first_name.capitalize()
-    prev_last_name = input("Enter previous last name: ")
-    prev_last_name = prev_last_name.capitalize()
+    prev_data = get_data(session)
+    prev_first_name, prev_last_name = prev_data[0], prev_data[1]
 
     people = session.query(People)
     for person in people:
-        if person.first_name == prev_first_name:
-            if person.last_name == prev_last_name:
-                print("\nNow update the person with their new data.\n")
-                print()
+        if person.first_name == prev_first_name and person.last_name == prev_last_name:
+            print("Now update the person with their new data.")
 
-                updated_first_name = input("Enter updated first name: ")
-                updated_first_name = updated_first_name.capitalize()
-                updated_last_name = input("Enter updated last name: ")
-                updated_last_name = updated_last_name.capitalize()
-                updated_dob = input("Enter updated date of birth (yyyy-mm-dd): ")
+            updated_data = get_data(session)
+            updated_first_name, updated_last_name, updated_dob = updated_data[0], updated_data[1], updated_data[2]
 
-                person.first_name = updated_first_name
-                person.last_name = updated_last_name
-                person.dob = updated_dob
-                
-                print("Successfully updated")
-                print()
-                session.commit()
-                return
+            person.first_name = updated_first_name
+            person.last_name = updated_last_name
+            person.dob = updated_dob
+            
+            print("Successfully updated")
+            print("\n")
+            session.commit()
+            return
 
-    print("Could not update.")
-    print()
+    print(f"{prev_first_name} {prev_last_name} not in database...")
+    print("Update failed.")
+    print("\n")
 
 
 def delete(session:None) -> None:
-    print("\nTo delete a persons data, enter their first and last names.")
-    print()
+    print("To delete a persons data, enter their first and last names.")
 
-    delete_first_name = input("Enter first name: ")
-    delete_first_name = delete_first_name.capitalize()
-    delete_last_name = input("Enter last name: ")
-    delete_last_name = delete_last_name.capitalize()
-    print()
+    delete_data = get_data(session)
+    delete_first_name, delete_last_name = delete_data[0], delete_data[1]
 
     print("Are you sure you want to delete this data?")
     confirm = input("Enter y/n: ")
@@ -150,11 +163,18 @@ def delete(session:None) -> None:
     if confirm == "y":
         people = session.query(People)
         for person in people:
-            if person.first_name == delete_first_name:
-                if person.last_name == delete_last_name:
-                    session.delete(person)
-                    session.commit()
-                    return
+            if person.first_name == delete_first_name and person.last_name == delete_last_name:
+                print("Data deleted.")
+                print("\n")
+                session.delete(person)
+                session.commit()
+                return
+        print("Persons data does not exist")
+        print("\n")
+        return
+    else:
+        print("\n")
+        return
 
 
 def main() -> None:
